+++ app/Providers/AppServiceProvider.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace App\Providers;
 
-use Override;
+use Illuminate\Support\Facades\Event;
 use Illuminate\Support\ServiceProvider;
+use Override;
+use SocialiteProviders\Manager\SocialiteWasCalled;
+use SocialiteProviders\Strava\StravaExtendSocialite;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -23,6 +26,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        //
+        Event::listen(SocialiteWasCalled::class, StravaExtendSocialite::class);
     }
 }